requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
idna==3.3
Pillow==9.0.0
pipreqs==0.4.11
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
yarg==0.1.9
